add dump function

+++ src/build/build.sh
@@ -1,10 +1,15 @@
 #! /bin/sh
 
+if [ $# -ne 1 ]; then
+    echo "Wrong number of arguments!"
+    exit 1
+fi
+
 if [ $1 == 'clean' ];then
     rm -fr CMakeFiles cmake_install.cmake CMakeCache.txt Makefile test.out utils.out polygon-net
 elif [ $1 == 'build' ];then
     cmake .
     make
 else
-    echo 'Unknown ag'
+    echo 'Unknown argument'
 fi
